Move byte array functions to utils
